requirements_test.txt
pytest
pytest-django
selenium
pytest-selenium